// src/types/priorAuthStatus.ts
export enum PriorAuthStatus {
  PENDING_INSURANCE = 'Pending Insurance',
  PENDING_MEDICAL = 'Pending Medical',
  PROCESSING = 'Processing',
  SUBMITTED = 'Submitted',
  APPROVED = 'Approved',
  APPROVED_WITH_PHARMACY_BENEFITS = 'Approved with Pharmacy Benefits',
  DENIED = 'Denied',
  CANCELLED = 'Cancelled',
  STOPPED = 'Stopped',
}